LayoutTests/platform/glib/imported/w3c/web-platform-tests/fetch/api/cors/cors-preflight-status.any.worker-expected.txt

CONSOLE MESSAGE: Preflight response is not successful. Status code: 300
CONSOLE MESSAGE: Preflight response is not successful. Status code: 301
CONSOLE MESSAGE: Preflight response is not successful. Status code: 302
CONSOLE MESSAGE: Preflight response is not successful. Status code: 303
CONSOLE MESSAGE: Preflight response is not successful. Status code: 304
CONSOLE MESSAGE: Preflight response is not successful. Status code: 305
CONSOLE MESSAGE: Preflight response is not successful. Status code: 306
CONSOLE MESSAGE: Preflight response is not successful. Status code: 307
CONSOLE MESSAGE: Preflight response is not successful. Status code: 308
CONSOLE MESSAGE: Preflight response is not successful. Status code: 400
CONSOLE MESSAGE: Preflight response is not successful. Status code: 401
CONSOLE MESSAGE: Preflight response is not successful. Status code: 402
CONSOLE MESSAGE: Preflight response is not successful. Status code: 403
CONSOLE MESSAGE: Preflight response is not successful. Status code: 404
CONSOLE MESSAGE: Preflight response is not successful. Status code: 405
CONSOLE MESSAGE: Preflight response is not successful. Status code: 501
CONSOLE MESSAGE: Preflight response is not successful. Status code: 502
CONSOLE MESSAGE: Preflight response is not successful. Status code: 503
CONSOLE MESSAGE: Preflight response is not successful. Status code: 504
CONSOLE MESSAGE: Preflight response is not successful. Status code: 505

PASS Preflight answered with status 200
PASS Preflight answered with status 201
PASS Preflight answered with status 202
PASS Preflight answered with status 203
PASS Preflight answered with status 204
PASS Preflight answered with status 205
PASS Preflight answered with status 206
PASS Preflight answered with status 300
PASS Preflight answered with status 301
PASS Preflight answered with status 302
PASS Preflight answered with status 303
PASS Preflight answered with status 304
PASS Preflight answered with status 305
PASS Preflight answered with status 306
PASS Preflight answered with status 307
PASS Preflight answered with status 308
PASS Preflight answered with status 400
PASS Preflight answered with status 401
PASS Preflight answered with status 402
PASS Preflight answered with status 403
PASS Preflight answered with status 404
PASS Preflight answered with status 405
PASS Preflight answered with status 501
PASS Preflight answered with status 502
PASS Preflight answered with status 503
PASS Preflight answered with status 504
PASS Preflight answered with status 505
